move: chat message 관련 컴포넌트 파일 message 폴더 안으로 이동 (#47)
