+++ src/components/chat/chatRoom/ChatRoomPanel.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import { useParams } from 'next/navigation';
 import Header from '@/components/common/Header';
-import ChatRoom from './ChatRoom';
+import ChatRoom from './message/ChatRoom';
 import ChatInputBar from '@/components/chat/chatRoom/input/ChatInputBar';
 import { ChatMessagePropsType, ChatHeaderInfoPropsType } from '@/types/chat';
 import { getChatMessage, getChatHeaderInfo } from '@/api/chatApi';
